requirements.txt
streamlit>=1.36.0
langchain>=0.2.10
langchain-community>=0.2.10
langgraph>=0.2.14
chromadb>=0.5.5
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.2.2
pypdf>=4.2.0
python-docx>=1.1.2
python-pptx>=1.0.2
pandas>=2.2.2
beautifulsoup4>=4.12.3
markdown>=3.6
python-dotenv>=1.0.1
psutil>=5.9.8
pydantic>=2.7.4
